+++ Product.Infrastructure/Repositories/RepositoryBase.cs
@@ -35,7 +35,7 @@ namespace Product.Infrastructure.Repositories
             return await _dbContext.Set<T>().ToListAsync();
         }
 
-        public async Task<T> GetByIdAsync(long id)
+        public async Task<T> GetByIdAsync(int id)
         {
             return await _dbContext.Set<T>().FindAsync(id);
         }
